first own changes to pysotp

+++ can_wrap_src/can_wrap.h
@@ -14,10 +14,10 @@ int raw_can_init(char *iface, unsigned int len);
 int raw_can_map_channel(uint32_t can_id);
 
 /* sends some data over raw can */
-int raw_can_send(unsigned int channel, char *data, unsigned int len);
+long raw_can_send(unsigned int channel, char *data, unsigned int len);
 
 /* receive data from socket*/
-int raw_can_receive(unsigned int channel, char *data, unsigned int *len, long int timeout);
+long raw_can_receive(unsigned int channel, char *data, unsigned int *len, long int timeout);
 
 /* empties the receive queue */
 int raw_can_flush_rx(unsigned int channel);
@@ -37,7 +37,7 @@ int iso_tp_map_channel(uint32_t tx_id, uint32_t rx_id);
 int iso_tp_send(unsigned int channel, char *data, unsigned int len);
 
 /* receive data from socket*/
-int iso_tp_receive(unsigned int channel, char *data, unsigned int *len, long int timeout);
+long iso_tp_receive(unsigned int channel, char *data, unsigned int *len, long int timeout);
 
 /* emties the receive queue */
 int iso_tp_flush_rx(unsigned int channel);
